Move Subscription interface closer to use-site
