+++ src/main/kotlin/no/njoh/pulseengine/core/console/Subscription.kt
@@ -1,0 +1,5 @@
+package no.njoh.pulseengine.core.console
+
+interface Subscription {
+    fun unsubscribe()
+}
