--- src/main/kotlin/no/njoh/pulseengine/core/shared/primitives/Subscription.kt
@@ -1,5 +1,0 @@
-package no.njoh.pulseengine.core.shared.primitives
-
-interface Subscription {
-    fun unsubscribe()
-}
